Add refactor and examples
