+++ data_generator/type_generator/numeric_generator.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from src.data_generator.config import Config
+from data_generator.config import Config
 
 class BaseType:
