+++ examples/__init__.py
@@ -1,0 +1,3 @@
+import sys
+
+sys.path.append('..')
